backend/requirements.txt
fastapi
uvicorn
pydantic
sse-starlette
python-dotenv
google-generativeai
chromadb
pymupdf
google-search-results